added search like query

+++ fantasy_air/db/migrate/20171114190927_create_players.rb
@@ -13,11 +13,11 @@ class CreatePlayers < ActiveRecord::Migration[5.1]
       t.integer :air_yards
       t.integer :tm_att
       t.integer :tm_airyards
-      t.integer :aypt
-      t.integer :racr
-      t.integer :ms_air_yards 
-      t.integer :target_share
-      t.integer :wopr
+      t.float :aypt
+      t.float :racr
+      t.float :ms_air_yards 
+      t.float :target_share
+      t.float :wopr
 
       t.timestamps
     end
